Rename automatic_assessment to validation
