+++ tor/validation/helpers.py
